backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0

# Gemini official client (or google-generativeai older)
google-genai>=1.0.0

gunicorn==20.1.0
